fix(types): FluidEvent assignment issues
The `observerAdded` and `observerRemoved` method signatures are now looser, which avoids assignment issues
when passing a FluidValue with narrowed events to a more generic function.

The once private properties of `FluidValue` are now "protected" so the event type can still be
inferred from its `E` type parameter.

The `FluidObserver` type is now looser in that its `parent` property can be any object type.
This allows compatibility with any object that had `setFluidGetter` called on it, instead of only `FluidValue` instances.

The `AnyEvent` type is no longer internal and now named `UnknownFluidEvent`. Similarly, the `UnsafeFluidEvent` type
is not @internal anymore.